services/core/MasterDriverAgent/master_driver/interfaces/modbus_tk/requirements.txt
modbus-tk==0.5.8
pyserial==3.1.1
PyYAML==3.12
subprocess32==3.2.7